rename recognizer.py and resolver.py to base.py and update imports
