+++ geoparser/modules/recognizers/__init__.py
@@ -1,0 +1,1 @@
+from geoparser.modules.recognizers.base import Recognizer
