+++ geoparser/modules/recognizers/base.py
