+++ geoparser/modules/resolvers/__init__.py
@@ -1,0 +1,1 @@
+from geoparser.modules.resolvers.base import Resolver
