+++ geoparser/modules/resolvers/base.py
